refactor: decompose ops/doc/mod.rs into focused submodules (#898)
Extract three self-contained sections from the 2325-line mod.rs:

- toml_preserve.rs (148 lines): apply_value_diff, json_to_toml_value,
  json_to_toml_item. TOML comment-preserving helpers with zero coupling
  to YAML code.

- yaml_cst.rs (212 lines): apply_yaml_mapping_diff, apply_yaml_sequence_diff,
  apply_yaml_sequence_resize, try_remove_subsequence, json_to_yaml_node,
  json_to_yaml_mapping. YAML CST comment-preserving helpers.

- navigate.rs (284 lines): navigate_mut, set_at_path, delete_at_selector,
  delete_where, move_at_path, deep_merge, update_matching.
  Format-agnostic JSON document navigation and mutation primitives.

mod.rs re-exports all public items so external API is unchanged.
mod.rs reduced from 2325 to 1678 lines (644 lines extracted).
