+++ src/ops/doc/mod.rs
@@ -1,8 +1,18 @@
 use crate::selector;
 use std::path::Path;
 
+mod navigate;
 mod preserve;
+mod toml_preserve;
+mod yaml_cst;
 mod yaml_splice;
+
+pub use navigate::{
+    deep_merge, delete_at_selector, delete_where, move_at_path, navigate_mut, set_at_path,
+    update_matching,
+};
+use toml_preserve::apply_value_diff;
+use yaml_cst::{apply_yaml_mapping_diff, apply_yaml_sequence_diff, try_remove_subsequence};
 
 pub use yaml_splice::needs_yaml_quoting;
 
@@ -180,380 +190,6 @@ fn try_preserve_yaml_array(
     Ok(None)
 }
 
-// -----------------------------------------------------------------------
-// TOML comment-preserving helpers
-// -----------------------------------------------------------------------
-
-/// Recursively walk `old` and `new` JSON value trees and apply only the
-/// differences to the `toml_edit::Item`, preserving comments and formatting
-/// on unchanged parts.
-fn apply_value_diff(item: &mut toml_edit::Item, old: &serde_json::Value, new: &serde_json::Value) {
-    if old == new {
-        return;
-    }
-
-    match (old, new) {
-        (serde_json::Value::Object(old_map), serde_json::Value::Object(new_map)) => {
-            // Try to get a mutable table reference from the item.
-            let table = if let Some(t) = item.as_table_mut() {
-                t
-            } else if item.as_inline_table_mut().is_some() {
-                // Inline table: fall back to wholesale replacement since
-                // inline tables don't carry per-key comments.
-                *item = json_to_toml_item(new);
-                return;
-            } else {
-                *item = json_to_toml_item(new);
-                return;
-            };
-
-            // Remove keys that no longer exist.
-            let removed: Vec<String> = old_map
-                .keys()
-                .filter(|k| !new_map.contains_key(k.as_str()))
-                .cloned()
-                .collect();
-            for k in &removed {
-                table.remove(k);
-            }
-
-            // Add new keys or recurse into changed values.
-            for (key, new_val) in new_map {
-                if let Some(old_val) = old_map.get(key) {
-                    if old_val != new_val
-                        && let Some(child) = table.get_mut(key)
-                    {
-                        apply_value_diff(child, old_val, new_val);
-                    }
-                } else {
-                    table.insert(key, json_to_toml_item(new_val));
-                }
-            }
-        }
-
-        (serde_json::Value::Array(old_arr), serde_json::Value::Array(new_arr))
-            if old_arr.len() == new_arr.len() =>
-        {
-            // Same-length arrays: recurse element by element.
-            if let Some(arr) = item.as_array_mut() {
-                for (i, (o, n)) in old_arr.iter().zip(new_arr.iter()).enumerate() {
-                    if o != n
-                        && let Some(v) = arr.get_mut(i)
-                    {
-                        *v = json_to_toml_value(n);
-                    }
-                }
-            } else if let Some(aot) = item.as_array_of_tables_mut() {
-                for (i, (o, n)) in old_arr.iter().zip(new_arr.iter()).enumerate() {
-                    if o != n
-                        && let Some(table_item) = aot.get_mut(i)
-                    {
-                        let mut tbl_item = toml_edit::Item::Table(table_item.clone());
-                        apply_value_diff(&mut tbl_item, o, n);
-                        if let toml_edit::Item::Table(t) = tbl_item {
-                            *table_item = t;
-                        }
-                    }
-                }
-            } else {
-                *item = json_to_toml_item(new);
-            }
-        }
-
-        // Type changed, different-length arrays, or scalar change:
-        // wholesale replacement.
-        _ => {
-            *item = json_to_toml_item(new);
-        }
-    }
-}
-
-/// Convert a `serde_json::Value` to a `toml_edit::Value` (scalar/array/inline-table).
-fn json_to_toml_value(val: &serde_json::Value) -> toml_edit::Value {
-    match val {
-        serde_json::Value::String(s) => toml_edit::Value::from(s.as_str()),
-        serde_json::Value::Bool(b) => toml_edit::Value::from(*b),
-        serde_json::Value::Number(n) => {
-            if let Some(i) = n.as_i64() {
-                toml_edit::Value::from(i)
-            } else {
-                // Covers u64 > i64::MAX and float values.
-                toml_edit::Value::from(n.as_f64().unwrap_or(0.0))
-            }
-        }
-        serde_json::Value::Array(arr) => {
-            let mut a = toml_edit::Array::new();
-            for v in arr {
-                a.push(json_to_toml_value(v));
-            }
-            toml_edit::Value::Array(a)
-        }
-        serde_json::Value::Object(map) => {
-            let mut t = toml_edit::InlineTable::new();
-            for (k, v) in map {
-                t.insert(k, json_to_toml_value(v));
-            }
-            toml_edit::Value::InlineTable(t)
-        }
-        serde_json::Value::Null => {
-            // TOML has no null; use empty string as fallback.
-            toml_edit::Value::from("")
-        }
-    }
-}
-
-/// Convert a `serde_json::Value` to a `toml_edit::Item`.
-///
-/// Objects become full `Table`s (not inline tables) so they render as
-/// `[section]` blocks. Arrays of objects become arrays-of-tables.
-fn json_to_toml_item(val: &serde_json::Value) -> toml_edit::Item {
-    match val {
-        serde_json::Value::Object(map) => {
-            let mut table = toml_edit::Table::new();
-            for (k, v) in map {
-                table.insert(k, json_to_toml_item(v));
-            }
-            toml_edit::Item::Table(table)
-        }
-        serde_json::Value::Array(arr) if !arr.is_empty() && arr.iter().all(|v| v.is_object()) => {
-            let mut aot = toml_edit::ArrayOfTables::new();
-            for v in arr {
-                if let serde_json::Value::Object(map) = v {
-                    let mut table = toml_edit::Table::new();
-                    for (k, v2) in map {
-                        table.insert(k, json_to_toml_item(v2));
-                    }
-                    aot.push(table);
-                }
-            }
-            toml_edit::Item::ArrayOfTables(aot)
-        }
-        _ => toml_edit::Item::Value(json_to_toml_value(val)),
-    }
-}
-
-// -----------------------------------------------------------------------
-// YAML comment-preserving helpers
-// -----------------------------------------------------------------------
-
-/// Recursively walk `old` and `new` JSON value trees and apply only the
-/// differences to the `yaml_edit::Mapping`, preserving comments and
-/// formatting on unchanged parts.
-/// Returns `Ok(true)` if all CST changes were fully applied, `Ok(false)`
-/// if at least one array resize was too complex for the CST path.
-fn apply_yaml_mapping_diff(
-    mapping: &yaml_edit::Mapping,
-    old: &serde_json::Value,
-    new: &serde_json::Value,
-) -> anyhow::Result<bool> {
-    if old == new {
-        return Ok(true);
-    }
-
-    let (Some(old_map), Some(new_map)) = (old.as_object(), new.as_object()) else {
-        return Ok(true);
-    };
-    let mut all_applied = true;
-
-    // Remove keys that no longer exist.
-    let removed: Vec<String> = old_map
-        .keys()
-        .filter(|k| !new_map.contains_key(k.as_str()))
-        .cloned()
-        .collect();
-    for k in &removed {
-        mapping.remove(k.as_str());
-    }
-
-    // Add new keys or recurse into changed values.
-    for (key, new_val) in new_map {
-        if let Some(old_val) = old_map.get(key) {
-            if old_val == new_val {
-                continue;
-            }
-            match (old_val, new_val) {
-                // Both objects: recurse using child view from get_mapping. Updates to
-                // pre-existing keys inside the sub use in-place set_value (preserves
-                // sibling inline comments). Brand-new keys inside may not attach on the
-                // cloned sub view (structure check will catch and fallback with comments
-                // preserved).
-                (serde_json::Value::Object(_), serde_json::Value::Object(_)) => {
-                    if let Some(child) = mapping.get_mapping(key.as_str()) {
-                        if !apply_yaml_mapping_diff(&child, old_val, new_val)? {
-                            all_applied = false;
-                        }
-                    } else {
-                        mapping.set(key.as_str(), json_to_yaml_mapping(new_val)?);
-                    }
-                }
-                // Both arrays: update via the existing sequence node to
-                // preserve block/flow style and the key-value newline.
-                (serde_json::Value::Array(old_arr), serde_json::Value::Array(new_arr)) => {
-                    if let Some(seq) = mapping.get_sequence(key.as_str()) {
-                        if old_arr.len() == new_arr.len() {
-                            if !apply_yaml_sequence_diff(&seq, old_arr, new_arr)? {
-                                all_applied = false;
-                            }
-                        } else if !apply_yaml_sequence_resize(
-                            &seq,
-                            old_arr,
-                            new_arr,
-                            mapping,
-                            key.as_str(),
-                            new_val,
-                        ) {
-                            all_applied = false;
-                        }
-                    } else {
-                        mapping.set(key.as_str(), json_to_yaml_node(new_val)?);
-                    }
-                }
-                // Type changed or scalar change.
-                _ => {
-                    mapping.set(key.as_str(), json_to_yaml_node(new_val)?);
-                }
-            }
-        } else {
-            // New key: add it.
-            if new_val.is_object() {
-                // Follow yaml-edit's own pattern for creating nested: set empty first,
-                // re-fetch the nested view (to get linked node), then populate.
-                // This ensures correct block indentation and attachment in the CST.
-                let empty = yaml_edit::Mapping::new();
-                mapping.set(key.as_str(), &empty);
-                if let Some(nested) = mapping.get_mapping(key.as_str()) {
-                    if let Some(obj) = new_val.as_object() {
-                        for (k, v) in obj {
-                            nested.set(k.as_str(), json_to_yaml_node(v)?);
-                        }
-                    }
-                } else {
-                    mapping.set(key.as_str(), json_to_yaml_mapping(new_val)?);
-                }
-            } else {
-                mapping.set(key.as_str(), json_to_yaml_node(new_val)?);
-            }
-        }
-    }
-    Ok(all_applied)
-}
-
-/// Element-by-element diff for same-length YAML sequences.
-/// Returns `Ok(true)` if all CST changes were fully applied.
-fn apply_yaml_sequence_diff(
-    seq: &yaml_edit::Sequence,
-    old_arr: &[serde_json::Value],
-    new_arr: &[serde_json::Value],
-) -> anyhow::Result<bool> {
-    let mut all_applied = true;
-    for (i, (o, n)) in old_arr.iter().zip(new_arr.iter()).enumerate() {
-        if o == n {
-            continue;
-        }
-        match (o, n) {
-            (serde_json::Value::Object(_), serde_json::Value::Object(_)) => {
-                if let Some(node) = seq.get(i)
-                    && let Some(child_mapping) = node.as_mapping()
-                {
-                    if !apply_yaml_mapping_diff(child_mapping, o, n)? {
-                        all_applied = false;
-                    }
-                    continue;
-                }
-                seq.set(i, json_to_yaml_node(n)?);
-            }
-            _ => {
-                seq.set(i, json_to_yaml_node(n)?);
-            }
-        }
-    }
-    Ok(all_applied)
-}
-
-/// Handle different-length array diffs while preserving comments.
-///
-/// Deletion is handled via targeted `Sequence::remove()` calls.
-/// Growth (prepend, append, general restructuring) leaves the CST
-/// unchanged; the caller in `serialize_value_preserving` handles it
-/// via text-level splicing so comments are always preserved.
-/// Returns `true` if the CST was successfully updated, `false` if the
-/// change was too complex for the CST path (caller should flag the result
-/// as needing a fallback).
-fn apply_yaml_sequence_resize(
-    seq: &yaml_edit::Sequence,
-    old_arr: &[serde_json::Value],
-    new_arr: &[serde_json::Value],
-    _mapping: &yaml_edit::Mapping,
-    _key: &str,
-    _new_val: &serde_json::Value,
-) -> bool {
-    if new_arr.len() < old_arr.len() && try_remove_subsequence(seq, old_arr, new_arr) {
-        return true;
-    }
-    // Growth or complex deletion: CST unchanged. Return false so the
-    // caller knows a text-level fallback is needed.
-    false
-}
-
-/// Try to remove elements from `seq` so that it matches `new_arr`,
-/// treating `new_arr` as an ordered subsequence of `old_arr`.
-///
-/// Returns `true` if the removal succeeded, `false` if `new_arr` is
-/// not a subsequence of `old_arr` (caller should fall back).
-fn try_remove_subsequence(
-    seq: &yaml_edit::Sequence,
-    old_arr: &[serde_json::Value],
-    new_arr: &[serde_json::Value],
-) -> bool {
-    let new_len = new_arr.len();
-    let mut remove_indices = Vec::new();
-    let mut ni = 0;
-    for (oi, old_item) in old_arr.iter().enumerate() {
-        if ni < new_len && *old_item == new_arr[ni] {
-            ni += 1;
-        } else {
-            remove_indices.push(oi);
-        }
-    }
-    if ni != new_len {
-        return false;
-    }
-    // Iterate in reverse to keep indices stable during removal.
-    for &idx in remove_indices.iter().rev() {
-        seq.remove(idx);
-    }
-    true
-}
-
-/// Convert a `serde_json::Value` to a `yaml_edit::YamlNode` by
-/// round-tripping through `serde_yaml_ng` (for correct serialization)
-/// and `yaml_edit` (for a CST node that `Mapping::set` can accept).
-///
-/// The value is embedded under a temporary key `__v__` so that
-/// `serde_yaml_ng` handles indentation of block sequences/mappings.
-fn json_to_yaml_node(val: &serde_json::Value) -> anyhow::Result<yaml_edit::YamlNode> {
-    use std::str::FromStr;
-    let wrapper = serde_json::json!({ "__v__": val });
-    let yaml_text = serde_yaml_ng::to_string(&wrapper)
-        .map_err(|e| anyhow::anyhow!("YAML serialization failed: {e}"))?;
-    let doc = yaml_edit::Document::from_str(&yaml_text)
-        .map_err(|e| anyhow::anyhow!("YAML CST re-parse failed: {e}"))?;
-    doc.as_mapping()
-        .and_then(|m| m.get("__v__"))
-        .ok_or_else(|| anyhow::anyhow!("YAML CST wrapper key missing"))
-}
-
-/// Convert a JSON object to a `yaml_edit::Mapping`.
-fn json_to_yaml_mapping(val: &serde_json::Value) -> anyhow::Result<yaml_edit::Mapping> {
-    let mapping = yaml_edit::Mapping::new();
-    if let Some(obj) = val.as_object() {
-        for (k, v) in obj {
-            mapping.set(k.as_str(), json_to_yaml_node(v)?);
-        }
-    }
-    Ok(mapping)
-}
-
 pub fn parse_doc(content: &str, format: &FileFormat) -> anyhow::Result<serde_json::Value> {
     match format {
         FileFormat::Json => Ok(serde_json::from_str(content)?),
@@ -611,289 +247,6 @@ fn resolve_yaml_merge_keys(value: &mut serde_json::Value) {
             }
         }
         _ => {}
-    }
-}
-
-pub fn navigate_mut<'a>(
-    root: &'a mut serde_json::Value,
-    segments: &[selector::Segment],
-    create: bool,
-) -> anyhow::Result<&'a mut serde_json::Value> {
-    let mut current = root;
-    for seg in segments {
-        current = match seg {
-            selector::Segment::Key(k) => {
-                if create {
-                    let needs_create = match current.as_object() {
-                        Some(obj) => !obj.contains_key(k.as_str()),
-                        None => false,
-                    };
-                    if needs_create {
-                        current
-                            .as_object_mut()
-                            .ok_or_else(|| anyhow::anyhow!("not an object at key '{k}'"))?
-                            .insert(k.clone(), serde_json::Value::Object(serde_json::Map::new()));
-                    }
-                }
-                current
-                    .get_mut(k.as_str())
-                    .ok_or_else(|| anyhow::anyhow!("key not found: {k}"))?
-            }
-            selector::Segment::Index(i) => current
-                .get_mut(*i)
-                .ok_or_else(|| anyhow::anyhow!("index out of bounds: {i}"))?,
-            _ => anyhow::bail!("wildcard/predicate not supported in write navigation"),
-        };
-    }
-    Ok(current)
-}
-
-/// Returns the parent path and the final segment (for use by set/delete/move).
-fn split_last(segments: &[selector::Segment]) -> (&[selector::Segment], &selector::Segment) {
-    let (parent, last) = segments.split_at(segments.len() - 1);
-    (parent, &last[0])
-}
-
-/// Set a value at the location described by `segments`.  Navigates to the
-/// parent (creating intermediate keys when needed) and inserts the value at
-/// the final Key or Index segment.
-pub fn set_at_path(
-    root: &mut serde_json::Value,
-    segments: &[selector::Segment],
-    value: serde_json::Value,
-) -> anyhow::Result<()> {
-    if segments.is_empty() {
-        anyhow::bail!("empty selector");
-    }
-    let (parent_path, last) = split_last(segments);
-    let parent = navigate_mut(root, parent_path, true)?;
-
-    match last {
-        selector::Segment::Key(k) => {
-            parent
-                .as_object_mut()
-                .ok_or_else(|| anyhow::anyhow!("parent is not an object"))?
-                .insert(k.clone(), value);
-        }
-        selector::Segment::Index(i) => {
-            let arr = parent
-                .as_array_mut()
-                .ok_or_else(|| anyhow::anyhow!("parent is not an array"))?;
-            if *i < arr.len() {
-                arr[*i] = value;
-            } else {
-                anyhow::bail!("index {} out of bounds (len {})", i, arr.len());
-            }
-        }
-        _ => anyhow::bail!("cannot set at wildcard/predicate"),
-    }
-    Ok(())
-}
-
-/// Delete the value at the given selector path. Returns `true` if
-/// something was removed, `false` if the path did not exist.
-pub fn delete_at_selector(
-    root: &mut serde_json::Value,
-    segments: &[selector::Segment],
-) -> anyhow::Result<bool> {
-    if segments.is_empty() {
-        return Ok(false);
-    }
-    let (parent_path, last) = split_last(segments);
-    let parent = match navigate_mut(root, parent_path, false) {
-        Ok(p) => p,
-        Err(_) => return Ok(false),
-    };
-    match last {
-        selector::Segment::Key(k) => {
-            if let Some(obj) = parent.as_object_mut() {
-                Ok(obj.remove(k.as_str()).is_some())
-            } else {
-                Ok(false)
-            }
-        }
-        selector::Segment::Index(i) => {
-            if let Some(arr) = parent.as_array_mut() {
-                if *i < arr.len() {
-                    arr.remove(*i);
-                    Ok(true)
-                } else {
-                    Ok(false)
-                }
-            } else {
-                Ok(false)
-            }
-        }
-        _ => anyhow::bail!("cannot delete at wildcard/predicate"),
-    }
-}
-
-/// Parse a `key=value` predicate and remove matching items from the array
-/// at `segments`. Returns the number of items removed.
-pub fn delete_where(
-    root: &mut serde_json::Value,
-    segments: &[selector::Segment],
-    predicate: &str,
-) -> anyhow::Result<usize> {
-    let eq_pos = predicate
-        .find('=')
-        .ok_or_else(|| anyhow::anyhow!("predicate must be in key=value format"))?;
-    let pred_key = predicate[..eq_pos].trim();
-    if pred_key.is_empty() {
-        anyhow::bail!("predicate key is empty; expected key=value format");
-    }
-    let raw_val = &predicate[eq_pos + 1..];
-    if raw_val.starts_with('=') {
-        anyhow::bail!("predicate uses '==' but only '=' is supported; use key=value format");
-    }
-    let pred_val = raw_val.trim();
-
-    let target = navigate_mut(root, segments, false)?;
-    let arr = target
-        .as_array_mut()
-        .ok_or_else(|| anyhow::anyhow!("selector does not point to an array"))?;
-
-    let before_len = arr.len();
-    arr.retain(|item| {
-        item.get(pred_key)
-            .is_none_or(|field| !selector::value_matches_str(field, pred_val))
-    });
-    Ok(before_len - arr.len())
-}
-
-/// Move a value from one path to another within the same document.
-/// Removes the value at `from_segments` and inserts it at `to_segments`.
-pub fn move_at_path(
-    root: &mut serde_json::Value,
-    from_segments: &[selector::Segment],
-    to_segments: &[selector::Segment],
-) -> anyhow::Result<()> {
-    // Remove value at source path.
-    let removed = {
-        if from_segments.is_empty() {
-            anyhow::bail!("empty from selector");
-        }
-        let (parent_path, last) = split_last(from_segments);
-        let parent = navigate_mut(root, parent_path, false)?;
-        match last {
-            selector::Segment::Key(k) => parent
-                .as_object_mut()
-                .and_then(|obj| obj.remove(k.as_str()))
-                .ok_or_else(|| anyhow::anyhow!("source key '{k}' not found"))?,
-            selector::Segment::Index(i) => {
-                let arr = parent
-                    .as_array_mut()
-                    .ok_or_else(|| anyhow::anyhow!("source parent is not an array"))?;
-                if *i < arr.len() {
-                    arr.remove(*i)
-                } else {
-                    anyhow::bail!("source index {i} out of bounds");
-                }
-            }
-            _ => anyhow::bail!("cannot move from wildcard/predicate"),
-        }
-    };
-
-    // Insert at destination path.
-    let (parent_path, last) = split_last(to_segments);
-    let parent = navigate_mut(root, parent_path, true)?;
-    match last {
-        selector::Segment::Key(k) => {
-            parent
-                .as_object_mut()
-                .ok_or_else(|| anyhow::anyhow!("target parent is not an object"))?
-                .insert(k.clone(), removed);
-        }
-        selector::Segment::Index(i) => {
-            let arr = parent
-                .as_array_mut()
-                .ok_or_else(|| anyhow::anyhow!("target parent is not an array"))?;
-            if *i <= arr.len() {
-                arr.insert(*i, removed);
-            } else {
-                anyhow::bail!("target index {i} out of bounds");
-            }
-        }
-        _ => anyhow::bail!("cannot move to wildcard/predicate"),
-    }
-    Ok(())
-}
-
-const MAX_MERGE_DEPTH: usize = 128;
-
-pub fn deep_merge(base: &mut serde_json::Value, other: &serde_json::Value) {
-    deep_merge_inner(base, other, 0);
-}
-
-fn deep_merge_inner(base: &mut serde_json::Value, other: &serde_json::Value, depth: usize) {
-    if depth >= MAX_MERGE_DEPTH {
-        *base = other.clone();
-        return;
-    }
-    if let (Some(base_map), Some(other_map)) = (base.as_object_mut(), other.as_object()) {
-        for (key, value) in other_map {
-            let entry = base_map
-                .entry(key.clone())
-                .or_insert(serde_json::Value::Null);
-            deep_merge_inner(entry, value, depth + 1);
-        }
-    } else {
-        *base = other.clone();
-    }
-}
-
-pub fn update_matching(
-    value: &mut serde_json::Value,
-    segments: &[selector::Segment],
-    new_val: &serde_json::Value,
-) -> usize {
-    if segments.is_empty() {
-        *value = new_val.clone();
-        return 1;
-    }
-    let first = &segments[0];
-    let rest = &segments[1..];
-    match first {
-        selector::Segment::Key(k) => {
-            if let Some(child) = value.get_mut(k.as_str()) {
-                update_matching(child, rest, new_val)
-            } else {
-                0
-            }
-        }
-        selector::Segment::Index(i) => {
-            if let Some(child) = value.get_mut(*i) {
-                update_matching(child, rest, new_val)
-            } else {
-                0
-            }
-        }
-        selector::Segment::Wildcard => {
-            let mut count = 0;
-            if let Some(arr) = value.as_array_mut() {
-                for item in arr.iter_mut() {
-                    count += update_matching(item, rest, new_val);
-                }
-            }
-            count
-        }
-        selector::Segment::Predicate {
-            key,
-            value: pred_val,
-        } => {
-            let mut count = 0;
-            if let Some(arr) = value.as_array_mut() {
-                for item in arr.iter_mut() {
-                    let matches = item
-                        .get(key.as_str())
-                        .is_some_and(|field| selector::value_matches_str(field, pred_val));
-                    if matches {
-                        count += update_matching(item, rest, new_val);
-                    }
-                }
-            }
-            count
-        }
     }
 }
