+++ src/ops/doc/yaml_cst.rs
@@ -1,0 +1,212 @@
+pub(crate) fn apply_yaml_mapping_diff(
+    mapping: &yaml_edit::Mapping,
+    old: &serde_json::Value,
+    new: &serde_json::Value,
+) -> anyhow::Result<bool> {
+    if old == new {
+        return Ok(true);
+    }
+
+    let (Some(old_map), Some(new_map)) = (old.as_object(), new.as_object()) else {
+        return Ok(true);
+    };
+    let mut all_applied = true;
+
+    // Remove keys that no longer exist.
+    let removed: Vec<String> = old_map
+        .keys()
+        .filter(|k| !new_map.contains_key(k.as_str()))
+        .cloned()
+        .collect();
+    for k in &removed {
+        mapping.remove(k.as_str());
+    }
+
+    // Add new keys or recurse into changed values.
+    for (key, new_val) in new_map {
+        if let Some(old_val) = old_map.get(key) {
+            if old_val == new_val {
+                continue;
+            }
+            match (old_val, new_val) {
+                // Both objects: recurse using child view from get_mapping. Updates to
+                // pre-existing keys inside the sub use in-place set_value (preserves
+                // sibling inline comments). Brand-new keys inside may not attach on the
+                // cloned sub view (structure check will catch and fallback with comments
+                // preserved).
+                (serde_json::Value::Object(_), serde_json::Value::Object(_)) => {
+                    if let Some(child) = mapping.get_mapping(key.as_str()) {
+                        if !apply_yaml_mapping_diff(&child, old_val, new_val)? {
+                            all_applied = false;
+                        }
+                    } else {
+                        mapping.set(key.as_str(), json_to_yaml_mapping(new_val)?);
+                    }
+                }
+                // Both arrays: update via the existing sequence node to
+                // preserve block/flow style and the key-value newline.
+                (serde_json::Value::Array(old_arr), serde_json::Value::Array(new_arr)) => {
+                    if let Some(seq) = mapping.get_sequence(key.as_str()) {
+                        if old_arr.len() == new_arr.len() {
+                            if !apply_yaml_sequence_diff(&seq, old_arr, new_arr)? {
+                                all_applied = false;
+                            }
+                        } else if !apply_yaml_sequence_resize(
+                            &seq,
+                            old_arr,
+                            new_arr,
+                            mapping,
+                            key.as_str(),
+                            new_val,
+                        ) {
+                            all_applied = false;
+                        }
+                    } else {
+                        mapping.set(key.as_str(), json_to_yaml_node(new_val)?);
+                    }
+                }
+                // Type changed or scalar change.
+                _ => {
+                    mapping.set(key.as_str(), json_to_yaml_node(new_val)?);
+                }
+            }
+        } else {
+            // New key: add it.
+            if new_val.is_object() {
+                // Follow yaml-edit's own pattern for creating nested: set empty first,
+                // re-fetch the nested view (to get linked node), then populate.
+                // This ensures correct block indentation and attachment in the CST.
+                let empty = yaml_edit::Mapping::new();
+                mapping.set(key.as_str(), &empty);
+                if let Some(nested) = mapping.get_mapping(key.as_str()) {
+                    if let Some(obj) = new_val.as_object() {
+                        for (k, v) in obj {
+                            nested.set(k.as_str(), json_to_yaml_node(v)?);
+                        }
+                    }
+                } else {
+                    mapping.set(key.as_str(), json_to_yaml_mapping(new_val)?);
+                }
+            } else {
+                mapping.set(key.as_str(), json_to_yaml_node(new_val)?);
+            }
+        }
+    }
+    Ok(all_applied)
+}
+
+/// Element-by-element diff for same-length YAML sequences.
+/// Returns `Ok(true)` if all CST changes were fully applied.
+pub(super) fn apply_yaml_sequence_diff(
+    seq: &yaml_edit::Sequence,
+    old_arr: &[serde_json::Value],
+    new_arr: &[serde_json::Value],
+) -> anyhow::Result<bool> {
+    let mut all_applied = true;
+    for (i, (o, n)) in old_arr.iter().zip(new_arr.iter()).enumerate() {
+        if o == n {
+            continue;
+        }
+        match (o, n) {
+            (serde_json::Value::Object(_), serde_json::Value::Object(_)) => {
+                if let Some(node) = seq.get(i)
+                    && let Some(child_mapping) = node.as_mapping()
+                {
+                    if !apply_yaml_mapping_diff(child_mapping, o, n)? {
+                        all_applied = false;
+                    }
+                    continue;
+                }
+                seq.set(i, json_to_yaml_node(n)?);
+            }
+            _ => {
+                seq.set(i, json_to_yaml_node(n)?);
+            }
+        }
+    }
+    Ok(all_applied)
+}
+
+/// Handle different-length array diffs while preserving comments.
+///
+/// Deletion is handled via targeted `Sequence::remove()` calls.
+/// Growth (prepend, append, general restructuring) leaves the CST
+/// unchanged; the caller in `serialize_value_preserving` handles it
+/// via text-level splicing so comments are always preserved.
+/// Returns `true` if the CST was successfully updated, `false` if the
+/// change was too complex for the CST path (caller should flag the result
+/// as needing a fallback).
+fn apply_yaml_sequence_resize(
+    seq: &yaml_edit::Sequence,
+    old_arr: &[serde_json::Value],
+    new_arr: &[serde_json::Value],
+    _mapping: &yaml_edit::Mapping,
+    _key: &str,
+    _new_val: &serde_json::Value,
+) -> bool {
+    if new_arr.len() < old_arr.len() && try_remove_subsequence(seq, old_arr, new_arr) {
+        return true;
+    }
+    // Growth or complex deletion: CST unchanged. Return false so the
+    // caller knows a text-level fallback is needed.
+    false
+}
+
+/// Try to remove elements from `seq` so that it matches `new_arr`,
+/// treating `new_arr` as an ordered subsequence of `old_arr`.
+///
+/// Returns `true` if the removal succeeded, `false` if `new_arr` is
+/// not a subsequence of `old_arr` (caller should fall back).
+pub(super) fn try_remove_subsequence(
+    seq: &yaml_edit::Sequence,
+    old_arr: &[serde_json::Value],
+    new_arr: &[serde_json::Value],
+) -> bool {
+    let new_len = new_arr.len();
+    let mut remove_indices = Vec::new();
+    let mut ni = 0;
+    for (oi, old_item) in old_arr.iter().enumerate() {
+        if ni < new_len && *old_item == new_arr[ni] {
+            ni += 1;
+        } else {
+            remove_indices.push(oi);
+        }
+    }
+    if ni != new_len {
+        return false;
+    }
+    // Iterate in reverse to keep indices stable during removal.
+    for &idx in remove_indices.iter().rev() {
+        seq.remove(idx);
+    }
+    true
+}
+
+/// Convert a `serde_json::Value` to a `yaml_edit::YamlNode` by
+/// round-tripping through `serde_yaml_ng` (for correct serialization)
+/// and `yaml_edit` (for a CST node that `Mapping::set` can accept).
+///
+/// The value is embedded under a temporary key `__v__` so that
+/// `serde_yaml_ng` handles indentation of block sequences/mappings.
+fn json_to_yaml_node(val: &serde_json::Value) -> anyhow::Result<yaml_edit::YamlNode> {
+    use std::str::FromStr;
+    let wrapper = serde_json::json!({ "__v__": val });
+    let yaml_text = serde_yaml_ng::to_string(&wrapper)
+        .map_err(|e| anyhow::anyhow!("YAML serialization failed: {e}"))?;
+    let doc = yaml_edit::Document::from_str(&yaml_text)
+        .map_err(|e| anyhow::anyhow!("YAML CST re-parse failed: {e}"))?;
+    doc.as_mapping()
+        .and_then(|m| m.get("__v__"))
+        .ok_or_else(|| anyhow::anyhow!("YAML CST wrapper key missing"))
+}
+
+/// Convert a JSON object to a `yaml_edit::Mapping`.
+fn json_to_yaml_mapping(val: &serde_json::Value) -> anyhow::Result<yaml_edit::Mapping> {
+    let mapping = yaml_edit::Mapping::new();
+    if let Some(obj) = val.as_object() {
+        for (k, v) in obj {
+            mapping.set(k.as_str(), json_to_yaml_node(v)?);
+        }
+    }
+    Ok(mapping)
+}
